+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 - registry: Log operating system quit signal as string
 
+### Added
+- registry/gc: Add Prometheus counter and histogram for online GC runs
+
 ## [v3.3.0-gitlab] - 2021-04-09
 ### Added
 - registry: Add Prometheus counter for database queries